fix bugs with nock interaction
1. do not stop instrumentation in MetricsTestHelper.afterEachTest()

If NewRelic.instrument() is triggered before nock is required and activated
then stopping instrumentation in afterEachTest() will prevent nock from
working, as the disabling restores the original http.request() method, thereby
removing the wrapper that nock added. All tests in downstream projects work
fine with instrumentation always on.

2. do not mess with the nock global scope

Use a separate metricsNock and only remove() that in afterEachTest()
instead of doing a cleanAll() that can interfere with other nocks in tests.